keeps underlines on definitions
