+++ crates/project/src/project.rs
@@ -4994,23 +4994,6 @@ impl Project {
         }
     }
 
-    pub fn exclude_link_to_position(
-        buffer: &Model<Buffer>,
-        snapshot: &BufferSnapshot,
-        current_position: &Anchor,
-        location: &LocationLink,
-    ) -> bool {
-        // Exclude definition links that points back to cursor position.
-        // (i.e., currently cursor upon definition).
-        !(buffer == &location.target.buffer
-            && current_position
-                .cmp(&location.target.range.start, &snapshot)
-                .is_ge()
-            && current_position
-                .cmp(&location.target.range.end, &snapshot)
-                .is_le())
-    }
-
     #[inline(never)]
     fn definition_impl(
         &self,
@@ -5025,7 +5008,6 @@ impl Project {
             cx,
         )
     }
-
     pub fn definition<T: ToPointUtf16>(
         &self,
         buffer: &Model<Buffer>,
